+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  * Add ability to name portions of a Trident computation and have those names appear in the Storm UI
  * Show Nimbus and topology configurations through Storm UI (thanks rnfein)
  * Added ITupleCollection interface for TridentState's and TupleCollectionGet QueryFunction for getting the full contents of a state. MemoryMapState and LRUMemoryMapState implement this
+ * Can now submit a topology in inactive state. Storm will wait to call open/prepare on the spouts/bolts until it is first activated.
  * Bug fix: Fix race condition in supervisor that would lead to supervisor continuously crashing due to not finding "stormconf.ser" file for an already killed topology
  
 ## 0.8.1